Move `SyntaxShape` specifier parsing into own file
